+++ src/services/index.ts
@@ -1,2 +1,3 @@
-export { generateRandomUniqueId, findUserPreferedVideo } from './user';
+export { generateRandomUniqueId } from './user';
 export { initKakao } from './kakao';
+export { findUserPreferedVideo } from './video';
